Remodularized MainGUI
Put objects in a seperate python file to avoid overwriting
